Refactor Hydro-MHD stuff into its own subcomponent
-> We needed to introduce an include statement into EnzoInitialBCenter.cpp since we reuse some of the Bfield functionality from the hydro-layer to help initialize magnetic fields
-> EnzoPermutedCoordinates got moved into the utils subdirectory because it can be used in more cases than just hydro. For example, it is used in the initializer for the Shock Tube test problem. But, I could easily be convinced to put it back.
