+++ src/Enzo/charm_enzo.cpp
@@ -27,6 +27,7 @@
 //   each other
 
 #include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 // next, make sure to include "charm_enzo.hpp" (which includes enzo.decl.h)
 #include "Enzo/charm_enzo.hpp"
