+++ src/Enzo/enzo-core/EnzoProblem.cpp
@@ -8,6 +8,8 @@
 ///
 
 #include "enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp" // EnzoMethodMHDVlct, EnzoMethodPpm,
+                                        // EnzoMethodPpml
 
 //----------------------------------------------------------------------
