+++ src/Enzo/hydro-mhd/EnzoMethodPpm.cpp
@@ -5,8 +5,9 @@
 /// @date     Fri Apr  2 17:05:23 PDT 2010
 /// @brief    Implements the EnzoMethodPpm class
 
-#include "cello.hpp"
-#include "enzo.hpp"
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 // #define DEBUG_PPM
 // #define COPY_FIELDS_TO_OUTPUT
