+++ src/Enzo/hydro-mhd/SetMinimumSupport.cpp
@@ -5,9 +5,9 @@
 /// @date      November, 1998
 /// @brief     Set the energy to provide minimal pressure support
 
-#include "cello.hpp"
-
-#include "enzo.hpp"
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
  
 //----------------------------------------------------------------------
