+++ src/Enzo/hydro-mhd/SolveMHDEquations.cpp
@@ -17,9 +17,9 @@
 
 // Solve the MHD equations with the solver, saving the subgrid fluxes
 
-#include "cello.hpp"
-
-#include "enzo.hpp"
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 int EnzoBlock::SolveMHDEquations( enzo_float dt )
 {
