+++ src/Enzo/hydro-mhd/toolkit/EnzoBfieldMethod.cpp
@@ -6,8 +6,9 @@
 /// @brief    [\ref Enzo] Implementation of the BfieldMethod abstract base
 ///           class.
 
-#include "cello.hpp"
-#include "enzo.hpp"
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 //----------------------------------------------------------------------
