+++ src/Enzo/hydro-mhd/toolkit/EnzoIntegrationQuanUpdate.cpp
@@ -5,9 +5,11 @@
 /// @date     Mon June 24 2019
 /// @brief    [\ref Enzo] Implementation of EnzoIntegrationQuanUpdate.
 
-#include "cello.hpp"
-#include "enzo.hpp"
 #include <limits>
+
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 //----------------------------------------------------------------------
