+++ src/Enzo/hydro-mhd/toolkit/EnzoLazyPassiveScalarFieldList.cpp
@@ -6,9 +6,9 @@
 /// @brief    [\ref Enzo] Implementation of EnzoLazyPassiveScalarFieldList
 ///           class
 
-#include "cello.hpp"
-#include "enzo.hpp"
-#include "charm_enzo.hpp"
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 //----------------------------------------------------------------------
