+++ src/Enzo/hydro-mhd/toolkit/EnzoSourceInternalEnergy.cpp
@@ -5,8 +5,9 @@
 /// @date     Fri December 13 2019
 /// @brief    [\ref Enzo] Implementation of Enzo's SourceInternalEnergy class.
 
-#include "cello.hpp"
-#include "enzo.hpp"
+#include "Cello/cello.hpp"
+#include "Enzo/enzo.hpp"
+#include "Enzo/hydro-mhd/hydro-mhd.hpp"
 
 //----------------------------------------------------------------------
